Properly split parity and parity-lib
